services/requirements.txt
pypdf>=3.0.0
langchain>=0.1.20
langchain-community>=0.0.38
groq>=0.4.1
langchain_groq>=0.1.4
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0
